Module: Fix test env usage example
Fixes the indentation issue in the module call, and assumes that Maker
is a constructor with a build method, which the tests then call.

Also, in the QUnit issue itself I suggested to fix this by deep-copying
the test environment properties. I've looked into that and concluded
that its way too complicated for something that no one ever had a problem
with, as far as I can tell. Instead we can just fix the example to use
a new parts array for each test.

Fixes #90
Fixes jquery/qunit#688